@@ -3,10 +3,11 @@ description: A Piwigo Android application
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 2.2.3+223
+version: 2.3.0+230
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: 3.22.2
 
 dependencies:
   flutter:
@@ -15,7 +16,7 @@ dependencies:
   # Network
   url_launcher: ^6.1.5 # Open links with device's applications
   webview_flutter: ^3.0.4 # Show web page (CGU)
-  dio: ^4.0.6 # Requests to API
+  dio: ^5.4.3+1 # Requests to API
   dio_cookie_manager: ^2.0.0 # Compatibility between dio and cookie_far
   cookie_jar: ^3.0.1 # Handles cookies
   connectivity_plus: ^3.0.2 # Check if Wifi is enabled
@@ -24,19 +25,22 @@ dependencies:
   auto_size_text: ^3.0.0 # Text that auto-sizes (image thumbnail title)
   cupertino_icons: ^1.0.2 # iOS style icons (might be useless)
   font_awesome_flutter: ^10.1.0 # Font awesome icons (might be useless)
-  flutter_slidable: ^2.0.0 # Album card's sliding cations
+  flutter_slidable: ^3.0.1 # Album card's sliding cations
   drag_select_grid_view: ^0.6.1 # Drag to select image grid
-  rounded_loading_button: ^2.1.0 # Loading button animation
+  rounded_loading_button: # Loading button animation
+    git:
+      url: https://github.com/scopendo/flutter_rounded_loading_button # A custom fork that can work with the latest flutter SDK
+      ref: dd4b76a
   modal_bottom_sheet: ^3.0.0-pre # Custom modals (might be useless)
   cached_network_image: ^3.2.2 # Better cache for images (used for album's thumbnail)
   flutter_speed_dial: ^6.1.0+1 # Speed dial
   pull_to_refresh: ^2.0.0 # Top and bottom refresh gestures
   photo_view: ^0.14.0 # Zoom on fullscreen photos
-  extended_text: ^11.0.0 # Text overflow on left side
+  extended_text: ^13.0.0 # Text overflow on left side
   flutter_easyloading: ^3.0.5 # Show loading dialog
 
   # Storage
-  package_info_plus: ^3.1.2 # Get project info (version)
+  package_info_plus: ^8.0.0 # Get project info (version)
   path_provider: ^2.0.11 # Get application documents
   shared_preferences: ^2.0.15 # Base local storage
   flutter_secure_storage: ^6.0.0 # Local storage secured for logins
@@ -44,19 +48,19 @@ dependencies:
   file_picker: ^5.2.5 # Pick images and videos at a time for upload
 
   # Device
-  device_info_plus: ^8.2.0 # Get device info (version)
-  flutter_local_notifications: ^12.0.2 # Throws notifications on download or upload
-  open_filex: ^4.3.2 # Open files with devices apps
+  device_info_plus: ^10.1.0 # Get device info (version)
+  flutter_local_notifications: ^17.2.0 # Throws notifications on download or upload
+  open_filex: ^4.4.0 # Open files with devices apps
   workmanager: ^0.5.0 # Background processes (auto upload)
   image_gallery_saver: ^2.0.3 # Download images
 
   # Utils
   mime_type: ^1.0.0 # Check mime type of files (differentiate photos from videos)
   video_player: ^2.4.7 # Read video files in fullscreen mode
-  chewie: ^1.5.0 # Video player with options
+  chewie: ^1.8.1 # Video player with options
   flutter_image_compress: ^1.1.3 # Remove metadata
   permission_handler: ^10.2.0 # Check and asks for permissions
-  share_plus: ^4.4.0 # Share files
+  share_plus: ^9.0.0 # Share files
   flutter_cache_manager: ^3.3.0 # Handles network image cache
   heic_to_jpg: ^0.2.0 # Convert heic files to jpg
   provider: ^6.0.3 # Notifiers for theme and language changes
@@ -68,7 +72,7 @@ dependencies:
   # Translations
   flutter_localizations:
     sdk: flutter
-  intl: ^0.18.0 # Used for translations
+  intl: ^0.19.0 # Used for translations
   html_unescape: ^2.0.0
 
 dev_dependencies:
