@@ -40,8 +40,7 @@ class AlbumPage extends StatefulWidget {
 }
 
 class _AlbumPageState extends State<AlbumPage> {
-  final RefreshController _refreshController =
-      RefreshController(initialRefresh: false);
+  final RefreshController _refreshController = RefreshController(initialRefresh: false);
   final ScrollController _scrollController = ScrollController();
 
   late AlbumModel _currentAlbum;
@@ -63,8 +62,7 @@ class _AlbumPageState extends State<AlbumPage> {
     super.initState();
   }
 
-  bool get _hasNonFavorites =>
-      _selectedList.where((image) => !image.favorite).isNotEmpty;
+  bool get _hasNonFavorites => _selectedList.where((image) => !image.favorite).isNotEmpty;
 
   bool get _enableLoad {
     if (_imageList == null || _imageList!.isEmpty) return false;
@@ -85,8 +83,7 @@ class _AlbumPageState extends State<AlbumPage> {
   Future<void> _loadMoreImages() async {
     if (_imageList == null) return;
     if (_currentAlbum.nbImages <= _imageList!.length) return;
-    ApiResponse<List<ImageModel>> result =
-        await fetchImages(widget.album.id, _page + 1);
+    ApiResponse<List<ImageModel>> result = await fetchImages(widget.album.id, _page + 1);
     if (result.hasError || !result.hasData) {
       _refreshController.loadFailed();
       await Future.delayed(const Duration(milliseconds: 500));
@@ -100,12 +97,9 @@ class _AlbumPageState extends State<AlbumPage> {
   }
 
   Future<void> _onRefresh() async {
-    final result = await Future.wait(
-        [fetchAlbums(widget.album.id), fetchImages(widget.album.id, 0)]);
-    final ApiResponse<List<AlbumModel>> albumsResult =
-        result.first as ApiResponse<List<AlbumModel>>;
-    final ApiResponse<List<ImageModel>> imagesResult =
-        result.last as ApiResponse<List<ImageModel>>;
+    final result = await Future.wait([fetchAlbums(widget.album.id), fetchImages(widget.album.id, 0)]);
+    final ApiResponse<List<AlbumModel>> albumsResult = result.first as ApiResponse<List<AlbumModel>>;
+    final ApiResponse<List<ImageModel>> imagesResult = result.last as ApiResponse<List<ImageModel>>;
     if (!albumsResult.hasData || !imagesResult.hasData) {
       _refreshController.refreshFailed();
       await Future.delayed(const Duration(milliseconds: 500));
@@ -120,14 +114,14 @@ class _AlbumPageState extends State<AlbumPage> {
     _refreshController.refreshCompleted();
   }
 
-  void _onAddAlbum() =>
-      onAddAlbum(context, widget.album.id).whenComplete(() => _onRefresh());
-  void _onTapAlbum(AlbumModel album) =>
-      onOpenAlbum(context, album).whenComplete(() => _onRefresh());
-  void _onEditAlbum(AlbumModel album) =>
-      onEditAlbum(context, album).whenComplete(() => _onRefresh());
-  void _onMoveAlbum(AlbumModel album) =>
-      onMoveAlbum(context, album).whenComplete(() => _onRefresh());
+  void _onAddAlbum() => onAddAlbum(context, widget.album.id).whenComplete(() => _onRefresh());
+
+  void _onTapAlbum(AlbumModel album) => onOpenAlbum(context, album).whenComplete(() => _onRefresh());
+
+  void _onEditAlbum(AlbumModel album) => onEditAlbum(context, album).whenComplete(() => _onRefresh());
+
+  void _onMoveAlbum(AlbumModel album) => onMoveAlbum(context, album).whenComplete(() => _onRefresh());
+
   Future<bool> _onDeleteAlbum(AlbumModel album) async {
     return onDeleteAlbum(context, album).then((success) {
       if (success) _onRefresh();
@@ -135,8 +129,7 @@ class _AlbumPageState extends State<AlbumPage> {
     });
   }
 
-  void _onAlbumPrivacy(AlbumModel album) =>
-      onEditAlbumPrivacy(context, album).whenComplete(() => _onRefresh());
+  void _onAlbumPrivacy(AlbumModel album) => onEditAlbumPrivacy(context, album).whenComplete(() => _onRefresh());
 
   void _onTapPhoto(ImageModel image) {
     Navigator.of(context).pushNamed(
@@ -165,13 +158,11 @@ class _AlbumPageState extends State<AlbumPage> {
   }
 
   Future<void> _onMovePhotos() async {
-    onMovePhotos(context, _selectedList, _currentAlbum)
-        .whenComplete(() => _onRefresh());
+    onMovePhotos(context, _selectedList, _currentAlbum).whenComplete(() => _onRefresh());
   }
 
   void _onLikePhotos() {
-    onLikePhotos(_selectedList, _hasNonFavorites)
-        .whenComplete(() => _onRefresh());
+    onLikePhotos(_selectedList, _hasNonFavorites).whenComplete(() => _onRefresh());
   }
 
   void _onDeletePhotos() {
@@ -206,100 +197,110 @@ class _AlbumPageState extends State<AlbumPage> {
     }).then((value) => _refreshController.requestRefresh());
   }
 
+  void _onWillPop(bool pop) {
+    if (pop) return;
+    if (_selectedList.isNotEmpty) {
+      setState(() {
+        _selectedList.clear();
+      });
+    } else {
+      Navigator.of(context).pop();
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: SmartRefresher(
-          controller: _refreshController,
-          scrollController: _scrollController,
-          enablePullUp: _enableLoad,
-          onLoading: _loadMoreImages,
-          onRefresh: _onRefresh,
-          header: MaterialClassicHeader(
-            backgroundColor: Theme.of(context).cardColor,
-            color: Theme.of(context).colorScheme.secondary,
-          ),
-          footer: ClassicFooter(
-            loadingText: appStrings.loadingHUD_label,
-            noDataText: appStrings.categoryImageList_noDataError,
-            failedText: appStrings.errorHUD_label,
-            idleText: '',
-            canLoadingText: appStrings.loadMoreHUD_label,
-          ),
-          child: CustomScrollView(
-            controller: _scrollController,
-            slivers: [
-              _appBar,
-              SliverToBoxAdapter(
-                child: FutureBuilder<List<ApiResponse>>(
-                  future: _data,
-                  builder: (context, snapshot) {
-                    if (snapshot.hasData) {
-                      if (snapshot.data!.first.hasError &&
-                          snapshot.data!.last.hasError) {
-                        return Padding(
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 16.0, vertical: 8.0),
-                          child: Text(
-                            appStrings.categoryImageList_noDataError,
-                            textAlign: TextAlign.center,
-                          ),
-                        );
-                      }
-                      return Column(
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        children: [
-                          _albumGrid(snapshot),
-                          _imageGrid(snapshot),
-                          SizedBox(
-                            height: 72.0,
-                            child: Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Text(
-                                appStrings
-                                    .imageCount(_currentAlbum.nbTotalImages),
-                                style: Theme.of(context).textTheme.titleSmall,
+    return PopScope(
+      canPop: false,
+      onPopInvoked: _onWillPop,
+      child: Scaffold(
+        body: SafeArea(
+          child: SmartRefresher(
+            controller: _refreshController,
+            scrollController: _scrollController,
+            enablePullUp: _enableLoad,
+            onLoading: _loadMoreImages,
+            onRefresh: _onRefresh,
+            header: MaterialClassicHeader(
+              backgroundColor: Theme.of(context).cardColor,
+              color: Theme.of(context).colorScheme.secondary,
+            ),
+            footer: ClassicFooter(
+              loadingText: appStrings.loadingHUD_label,
+              noDataText: appStrings.categoryImageList_noDataError,
+              failedText: appStrings.errorHUD_label,
+              idleText: '',
+              canLoadingText: appStrings.loadMoreHUD_label,
+            ),
+            child: CustomScrollView(
+              controller: _scrollController,
+              slivers: [
+                _appBar,
+                SliverToBoxAdapter(
+                  child: FutureBuilder<List<ApiResponse>>(
+                    future: _data,
+                    builder: (context, snapshot) {
+                      if (snapshot.hasData) {
+                        if (snapshot.data!.first.hasError && snapshot.data!.last.hasError) {
+                          return Padding(
+                            padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+                            child: Text(
+                              appStrings.categoryImageList_noDataError,
+                              textAlign: TextAlign.center,
+                            ),
+                          );
+                        }
+                        return Column(
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            _albumGrid(snapshot),
+                            _imageGrid(snapshot),
+                            SizedBox(
+                              height: 72.0,
+                              child: Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Text(
+                                  appStrings.imageCount(_currentAlbum.nbTotalImages),
+                                  style: Theme.of(context).textTheme.titleSmall,
+                                ),
                               ),
                             ),
-                          ),
-                        ],
+                          ],
+                        );
+                      }
+                      return const Center(
+                        child: Padding(
+                          padding: EdgeInsets.all(8.0),
+                          child: CircularProgressIndicator(),
+                        ),
                       );
-                    }
-                    return const Center(
-                      child: Padding(
-                        padding: EdgeInsets.all(8.0),
-                        child: CircularProgressIndicator(),
-                      ),
-                    );
-                  },
+                    },
+                  ),
                 ),
-              ),
-            ],
+              ],
+            ),
           ),
         ),
+        bottomNavigationBar: AnimatedSlide(
+          duration: const Duration(milliseconds: 300),
+          curve: Curves.easeInOut,
+          offset: _selectedList.isEmpty ? Offset(0, 1) : Offset.zero,
+          child: _bottomBar,
+        ),
+        floatingActionButton: _adminActionsSpeedDial,
       ),
-      bottomNavigationBar: AnimatedSlide(
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeInOut,
-        offset: _selectedList.isEmpty ? Offset(0, 1) : Offset.zero,
-        child: _bottomBar,
-      ),
-      floatingActionButton: _adminActionsSpeedDial,
     );
   }
 
   Widget get _appBar {
     Orientation orientation = MediaQuery.of(context).orientation;
     return SliverAppBar(
-      backgroundColor: Theme.of(context).colorScheme.background,
+      backgroundColor: Theme.of(context).colorScheme.surface,
       pinned: true,
       titleSpacing: 0,
       centerTitle: _selectedList.isNotEmpty,
       title: Text(
-        _selectedList.isEmpty
-            ? _currentAlbum.name
-            : _selectedList.length.toString(),
+        _selectedList.isEmpty ? _currentAlbum.name : _selectedList.length.toString(),
         style: Theme.of(context).appBarTheme.titleTextStyle,
       ),
       actions: [
@@ -311,8 +312,7 @@ class _AlbumPageState extends State<AlbumPage> {
             tooltip: appStrings.categoryImageList_deselectButton,
             icon: Icon(Icons.cancel),
           ),
-        if (orientation == Orientation.landscape && _selectedList.isNotEmpty)
-          ..._imageActions,
+        if (orientation == Orientation.landscape && _selectedList.isNotEmpty) ..._imageActions,
         if (widget.isAdmin || _currentAlbum.canUpload)
           PopupMenuButton(
             tooltip: appStrings.imageOptions_title,
@@ -329,17 +329,14 @@ class _AlbumPageState extends State<AlbumPage> {
                     text: appStrings.imageOptions_share,
                   ),
                 ),
-              if (_selectedList.isNotEmpty &&
-                  Preferences.getUserStatus != 'guest')
+              if (_selectedList.isNotEmpty && Preferences.getUserStatus != 'guest')
                 PopupMenuItem(
                   onTap: () => Future.delayed(
                     const Duration(seconds: 0),
                     _onLikePhotos,
                   ),
                   child: PopupListItem(
-                    icon: _hasNonFavorites
-                        ? Icons.favorite_border
-                        : Icons.favorite,
+                    icon: _hasNonFavorites ? Icons.favorite_border : Icons.favorite,
                     text: _hasNonFavorites
                         ? appStrings.imageOptions_addFavorites
                         : appStrings.imageOptions_removeFavorites,
@@ -400,8 +397,7 @@ class _AlbumPageState extends State<AlbumPage> {
   Widget _albumGrid(AsyncSnapshot snapshot) {
     // initialize album list
     if (_albumList == null) {
-      final ApiResponse<List<AlbumModel>> result =
-          snapshot.data!.first as ApiResponse<List<AlbumModel>>;
+      final ApiResponse<List<AlbumModel>> result = snapshot.data!.first as ApiResponse<List<AlbumModel>>;
       // if only albums has error
       if (!result.hasData) {
         return Center(
@@ -424,8 +420,7 @@ class _AlbumPageState extends State<AlbumPage> {
   Widget _imageGrid(AsyncSnapshot snapshot) {
     // Initialize image list
     if (_imageList == null) {
-      final ApiResponse<List<ImageModel>> result =
-          snapshot.data!.last as ApiResponse<List<ImageModel>>;
+      final ApiResponse<List<ImageModel>> result = snapshot.data!.last as ApiResponse<List<ImageModel>>;
       // if only images has error
       if (!result.hasData) {
         return Center(
@@ -434,13 +429,11 @@ class _AlbumPageState extends State<AlbumPage> {
       }
       _imageList = result.data!;
       // Refresh after build (for _enableLoad)
-      WidgetsBinding.instance
-          .addPostFrameCallback((timeStamp) => setState(() {}));
+      WidgetsBinding.instance.addPostFrameCallback((timeStamp) => setState(() {}));
     }
     if (_imageList!.isEmpty) return const SizedBox();
     // rebuild current selection with new images
-    _selectedList =
-        _imageList!.where((image) => _selectedList.contains(image)).toList();
+    _selectedList = _imageList!.where((image) => _selectedList.contains(image)).toList();
     return ImageGridView(
       album: _currentAlbum,
       imageList: _imageList!,
@@ -465,9 +458,7 @@ class _AlbumPageState extends State<AlbumPage> {
           builder: (context, uploadNotifier, child) {
             bool uploading = uploadNotifier.uploadList.isNotEmpty;
             return FloatingActionButton(
-              tooltip: uploading
-                  ? appStrings.uploadList_title
-                  : appStrings.categorySelection_root,
+              tooltip: uploading ? appStrings.uploadList_title : appStrings.categorySelection_root,
               shape: uploading ? CircleBorder() : null,
               backgroundColor: Theme.of(context).disabledColor.withOpacity(0.7),
               onPressed: () {
@@ -562,14 +553,11 @@ class _AlbumPageState extends State<AlbumPage> {
       return AnimatedContainer(
         duration: const Duration(milliseconds: 300),
         curve: Curves.easeInOut,
-        height: _selectedList.isEmpty || orientation == Orientation.landscape
-            ? 0
-            : 56.0,
+        height: _selectedList.isEmpty || orientation == Orientation.landscape ? 0 : 56.0,
         child: BottomAppBar(
           height: 56.0,
           child: Row(
-            children:
-                _imageActions.map((action) => Expanded(child: action)).toList(),
+            children: _imageActions.map((action) => Expanded(child: action)).toList(),
           ),
         ),
       );
@@ -603,9 +591,7 @@ class _AlbumPageState extends State<AlbumPage> {
       if (Preferences.getUserStatus != 'guest') // Todo: enum roles
         IconButton(
           onPressed: _onLikePhotos,
-          tooltip: _hasNonFavorites
-              ? appStrings.imageOptions_addFavorites
-              : appStrings.imageOptions_removeFavorites,
+          tooltip: _hasNonFavorites ? appStrings.imageOptions_addFavorites : appStrings.imageOptions_removeFavorites,
           isSelected: !_hasNonFavorites,
           selectedIcon: Icon(Icons.favorite),
           icon: Icon(Icons.favorite_border),
@@ -617,8 +603,6 @@ class _AlbumPageState extends State<AlbumPage> {
       ),
     ];
 
-    return widget.isAdmin || _currentAlbum.canUpload
-        ? adminActions
-        : userActions;
+    return widget.isAdmin || _currentAlbum.canUpload ? adminActions : userActions;
   }
 }
